requirements.txt
-e .
aws-cdk.aws_ec2